+++ android/app/src/main/java/net/melisma/selko/ui/theme/Theme.kt
@@ -1,10 +1,13 @@
 package net.melisma.selko.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Shapes
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.dp
 
 private val LightColorScheme = lightColorScheme(
     primary = SelkoPrimary,
@@ -44,6 +47,14 @@ private val DarkColorScheme = darkColorScheme(
     onError = SelkoOnError
 )
 
+private val SelkoShapes = Shapes(
+    extraSmall = RoundedCornerShape(2.dp),
+    small = RoundedCornerShape(2.dp),
+    medium = RoundedCornerShape(2.dp),
+    large = RoundedCornerShape(2.dp),
+    extraLarge = RoundedCornerShape(2.dp)
+)
+
 @Composable
 fun SelkoTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
@@ -54,6 +65,7 @@ fun SelkoTheme(
     MaterialTheme(
         colorScheme = colorScheme,
         typography = Typography,
+        shapes = SelkoShapes,
         content = content
     )
 }